Module Config.Ciphers

Cipher selection

Cipher selection

val default : Ciphersuite.ciphersuite list

default is a list of ciphersuites this library uses by default.

val supported : Ciphersuite.ciphersuite list

supported is a list of ciphersuites this library supports (larger than default).

val fs : Ciphersuite.ciphersuite list

fs is a list of ciphersuites which provide forward secrecy (sublist of default).

val fs_of : Ciphersuite.ciphersuite list -> Ciphersuite.ciphersuite list

fs_of ciphers selects all ciphersuites which provide forward secrecy from ciphers.